pyproject.toml
[project]
name = "feedback"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-vite>=3.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "nox>=2025.5.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.4",
]
